refactor(web): extract result-display components into components/results/ (issue #642, Phase 3 seam 4 pt 2)
Moves ResultPanel (+renderGroupLabel), SolveReadout (+feedMag,
GROUND_APPLIED_LABEL), ViewPanel, and PatternCompareTable out of App.tsx
into their own files under components/results/, mirroring the
components/charts/ and components/params/ split from earlier seams.

ViewPanel is the only one that reaches back into App.tsx for a type:
PatternData is a plain data shape (not tied to any component) that
App.tsx still uses elsewhere, so it stays put and ViewPanel takes a
type-only `import type { PatternData } from "../../App"` — build-time
erased, same precedent as ParamForm.tsx's KnobOpt import. The other
three files need nothing from App.tsx; their types all come from lib/
or components/charts/types.

Verbatim move (export added) — no behavior change. App.tsx keeps its
own imports of GHOST_COLOR_COUNT, SolveResponse/SolveRequest, and the
lib/view and lib/api types where DesignSession still uses them directly.
