+++ src/antennaknobs/web/frontend/src/App.tsx
@@ -1,6 +1,5 @@
 import {
   createContext,
-  Fragment,
   useCallback,
   useContext,
   useEffect,
@@ -37,11 +36,6 @@ import type {
   TerrainParams,
   TerrainPresetSchema,
 } from "./lib/ground";
-import {
-  formatOhms,
-  formatScalar,
-  formatSwr,
-} from "./lib/format";
 import { richardsonExtrap } from "./lib/math";
 import {
   familyOf,
@@ -55,8 +49,6 @@ import {
   type BandSpec,
   type ExampleDescriptor,
   type ParamValueBag,
-  type ResultFieldSpec,
-  type ResultSchemaItem,
   type SchemaItem,
   type SchemaParamGroupSpec,
   type SchemaParamSpec,
@@ -90,10 +82,7 @@ import {
   setCutsWsSend,
   type CutsWsMessage,
 } from "./components/charts/cuts";
-import { CurrentCanvas } from "./components/charts/CurrentCanvas";
-import { FarFieldChart } from "./components/charts/FarFieldChart";
-import { GHOST_COLOR_COUNT, GHOST_FALLBACK_RGB } from "./components/charts/palette";
-import { SmithChart } from "./components/charts/SmithChart";
+import { GHOST_COLOR_COUNT } from "./components/charts/palette";
 import type { PatternMetrics, PinnedPattern } from "./components/charts/types";
 import {
   ThemeContext,
@@ -103,6 +92,9 @@ import {
   useSlideSize,
   useThumbColumnSize,
 } from "./components/hooks";
+import { PatternCompareTable } from "./components/results/PatternCompareTable";
+import { SolveReadout } from "./components/results/SolveReadout";
+import { ViewPanel } from "./components/results/ViewPanel";
 
 // One advisory finding from the design screener (what a design does that a
 // typical one doesn't), attached to a trust-required entry.
@@ -118,72 +110,6 @@ export type DesignLoadError = {
   trust_required?: boolean;
   advisory?: DesignAdvisory[];
 };
-
-// label_template substitutions for ResultGroupItem:
-//   {i}            → 0-based index
-//   {i1}           → 1-based index
-//   {name:.Nf}     → result[name][i] formatted as a fixed-N-decimal float
-function renderGroupLabel(template: string, i: number, result: Record<string, unknown> | null): string {
-  let out = template.replace(/\{i1\}/g, String(i + 1)).replace(/\{i\}/g, String(i));
-  out = out.replace(/\{(\w+):\.(\d+)f\}/g, (_, name: string, decimals: string) => {
-    const arr = result?.[name];
-    if (!Array.isArray(arr)) return "—";
-    const v = arr[i];
-    return typeof v === "number" ? v.toFixed(Number(decimals)) : "—";
-  });
-  return out;
-}
-
-function ResultPanel({
-  schema,
-  result,
-}: {
-  schema: ResultSchemaItem[];
-  result: Record<string, unknown> | null;
-}) {
-  // Render one row per schema entry. Scalar items read the field off the
-  // response by name; group items repeat over the first inner field's
-  // top-level array. Missing/non-numeric values render as an em-dash so
-  // the row layout doesn't collapse mid-update.
-  return (
-    <>
-      {schema.map((item) => {
-        if ("kind" in item && item.kind === "group") {
-          const repeatField = item.fields[0]?.field;
-          const arr = repeatField ? result?.[repeatField] : undefined;
-          if (!Array.isArray(arr)) return null;
-          return (
-            <Fragment key={`result-group-${item.name}`}>
-              {arr.map((_, i) => (
-                <div className="row" key={`result-group-${item.name}-${i}`}>
-                  <span>{renderGroupLabel(item.label_template, i, result)}</span>
-                  <span className="val">
-                    {item.fields.map((f, fi) => {
-                      const sub = result?.[f.field];
-                      const v = Array.isArray(sub) ? sub[i] : undefined;
-                      return (
-                        <span key={`${item.name}-${i}-${fi}`}>
-                          {formatScalar(v, f.precision, f.unit)}
-                        </span>
-                      );
-                    })}
-                  </span>
-                </div>
-              ))}
-            </Fragment>
-          );
-        }
-        const s = item as ResultFieldSpec;
-        return (
-          <div className="row" key={`result-${s.field}`}>
-            <span>{s.label}</span>
-            <span className="val">{formatScalar(result?.[s.field], s.precision, s.unit)}</span>
-          </div>
-        );
-      })}
-    </>
-  );
-}
 
 // Log-spaced segments-per-wire ladder for the convergence sweep. Hentenna's
 // 8N+2 total segments at N=68 puts the dense LU at a ~550-cell matrix —
@@ -4246,361 +4172,6 @@ export function App() {
   );
 }
 
-function feedMag(r: SolveResponse): number {
-  const w = r.wires[r.feed_wire_index];
-  if (!w) return 0;
-  const re = w.knot_currents_re[r.feed_knot_index];
-  const im = w.knot_currents_im[r.feed_knot_index];
-  return Math.hypot(re, im);
-}
-
-// Display labels for SolveResponse.ground_model_applied — what the
-// impedance solve actually ran, as reported by the server (see the type).
-const GROUND_APPLIED_LABEL: Record<string, string> = {
-  sommerfeld: "Sommerfeld",
-  "refl-coef": "refl-coef",
-  "pec-image": "PEC image",
-  free: "free space",
-  // Faceted terrain: impedance ran crest-medium Sommerfeld; the far field
-  // reflects per-direction off the facets (issue #534).
-  terrain: "terrain (crest Somm.)",
-};
-
-// The R/X/SWR/rtt solve readout. The desktop stage floats it over the canvas
-// as a HUD (className="stage-readout"); the mobile Info screen (Phase B)
-// renders it as a normal block. Module scope so both trees share one
-// implementation.
-function SolveReadout({
-  result,
-  rttMs,
-  currentExample,
-  effectiveMultiFeed,
-  normCheck,
-  normCheckEnabled,
-  className = "",
-}: {
-  result: SolveResponse | null;
-  rttMs: number | null;
-  currentExample: ExampleDescriptor | undefined;
-  effectiveMultiFeed: boolean;
-  normCheck: NormCheckData | null;
-  normCheckEnabled: boolean;
-  className?: string;
-}) {
-  return (
-    <div className={`readout${className ? " " + className : ""}`}>
-      <div className="row">
-        <span>R</span>
-        <span className="val">{result ? formatOhms(result.z_in_re) : "—"}</span>
-      </div>
-      <div className="row">
-        <span>X</span>
-        <span
-          className={
-            result && Math.abs(result.z_in_im) < 2 && Math.abs(result.z_in_re) < 1e8
-              ? "val val-hot"
-              : "val"
-          }
-        >
-          {result
-            ? Math.abs(result.z_in_re) >= 1e8
-              ? "∞ (open)"
-              : formatOhms(result.z_in_im)
-            : "—"}
-        </span>
-      </div>
-      {currentExample && (
-        <ResultPanel
-          schema={currentExample.result_schema}
-          result={result as Record<string, unknown> | null}
-        />
-      )}
-      {effectiveMultiFeed && result?.feeds && result.feeds.length > 1 && (
-        <div className="feeds-table">
-          <div className="feeds-table-header">per-feed Z (V/I)</div>
-          {result.feeds.map((f, i) => (
-            <div className="row" key={`feed-z-${i}`}>
-              <span>
-                feed {i} ∠{Math.round(Math.atan2(f.v_im, f.v_re) * 180 / Math.PI)}°
-              </span>
-              <span className="val">
-                {f.z_re.toFixed(1)} {f.z_im >= 0 ? "+" : "−"} j
-                {Math.abs(f.z_im).toFixed(1)} Ω
-              </span>
-            </div>
-          ))}
-        </div>
-      )}
-      <div className="row">
-        <span>|I_feed|</span>
-        <span className="val">
-          {result ? feedMag(result).toExponential(3) : "—"}
-        </span>
-      </div>
-      {result?.ground && result.ground_model_applied && (
-        <div
-          className="row"
-          title="Ground model the impedance solve actually used, as reported by the solver — may be an approximation of the requested ground."
-        >
-          <span>ground</span>
-          <span className="val">
-            {GROUND_APPLIED_LABEL[result.ground_model_applied] ??
-              result.ground_model_applied}
-          </span>
-        </div>
-      )}
-      <div className="row">
-        <span>SWR ({(result?.z0_ohms ?? 50).toFixed(0)} Ω)</span>
-        <span className="val">
-          {result
-            ? formatSwr(result.z_in_re, result.z_in_im, result.z0_ohms ?? 50)
-            : "—"}
-        </span>
-      </div>
-      {(() => {
-        // Power budget (issue #299): where the input watts go, per network
-        // branch. Branch rows are hidden unless the network actually
-        // dissipates something (lossless branches report float noise only).
-        // The radiated row (issue #339) is the third efficiency ledger —
-        // P_radiated/P_input INCLUDING far-field ground absorption, derived
-        // from the dwell-triggered norm check — so it renders even for a
-        // lossless design over real ground, greyed to "—" while knobs move.
-        const budget = result?.power_budget;
-        const pin = result?.input_power_w;
-        const diss =
-          budget && pin ? budget.reduce((s, b) => s + b.watts, 0) : 0;
-        const showBudget =
-          !!budget && budget.length > 0 && !!pin && pin > 0 &&
-          diss >= 1e-6 * pin;
-        if (!showBudget && !normCheckEnabled) return null;
-        return (
-          <div className="feeds-table">
-            {showBudget && pin && (
-              <div title="Fraction of the source input power dissipated in each network branch (from the MNA solve); the antenna row is the remainder that reaches the wires.">
-                <div className="feeds-table-header">power budget</div>
-                {budget.map((b, i) => {
-                  // Hierarchical rows (issue #489): rows carry the instance
-                  // path of the composite they came from. Start a group
-                  // header whenever the path changes, and indent members
-                  // one step per hierarchy level ("sta.tuner" = depth 2).
-                  const path = b.path ?? "";
-                  const prev = i > 0 ? budget[i - 1].path ?? "" : "";
-                  const depth = path ? path.split(".").length : 0;
-                  return (
-                    <Fragment key={`pb-${i}`}>
-                      {path && path !== prev && (
-                        <div
-                          className="row pb-group"
-                          style={{ paddingLeft: `${(depth - 1) * 12}px` }}
-                        >
-                          <span>{path}</span>
-                        </div>
-                      )}
-                      <div
-                        className="row"
-                        style={
-                          depth ? { paddingLeft: `${depth * 12}px` } : undefined
-                        }
-                      >
-                        <span>{b.label}</span>
-                        <span className="val">
-                          {((b.watts / pin) * 100).toFixed(1)}%
-                        </span>
-                      </div>
-                    </Fragment>
-                  );
-                })}
-                <div className="row" key="pb-ant">
-                  <span>antenna (accepted)</span>
-                  <span className="val">
-                    {(((pin - diss) / pin) * 100).toFixed(1)}%
-                  </span>
-                </div>
-              </div>
-            )}
-            {normCheckEnabled &&
-              (() => {
-                // Over faceted terrain the same row reads the PEC-facet-
-                // referenced fraction (the tooltip explains the different
-                // denominator); the plain finite ground keeps its P_in-
-                // referenced number.
-                const isTerrain = !!normCheck?.method.startsWith("grid_terrain");
-                return (
-                  <div
-                    className="row"
-                    title={
-                      isTerrain
-                        ? "Share of accepted power leaving as sky wave over the faceted terrain: referenced against the same facet geometry with perfect-reflector media, so the ratio isolates real ground-media absorption (the hybrid model's separate field-vs-circuit ledger gap cancels; it is shown on the chart as the dotted overlay)."
-                        : "P_radiated / P_input from the dwell-triggered pattern integral (the norm check as a percentage): what actually leaves as far-field radiation after network, wire AND real ground absorption. Fills in once the knobs settle; over PEC ground or free space it collapses onto the structural efficiency. See the 'three ledgers' section of the docs."
-                    }
-                  >
-                    <span>radiated (incl. ground)</span>
-                    <span className={normCheck ? "val" : "val val-pending"}>
-                      {normCheck
-                        ? `${(normCheck.radiated_fraction * 100).toFixed(0)}%`
-                        : "—"}
-                    </span>
-                  </div>
-                );
-              })()}
-          </div>
-        );
-      })()}
-      {/* Engine timing grouped last: solve/rtt describe how fast the answer
-          arrived, not what the antenna is doing, so they sit below the RF
-          readout (and below the power budget when one is shown). The
-          feeds-table wrapper is used only for its dashed separator rule —
-          no header. */}
-      <div className="feeds-table">
-        {result?.solver_diag && (
-          <div
-            className="row"
-            title={
-              result.solver_diag.reason
-                ? `Array Block solve did not use the fast lattice-FFT coupling path: ${result.solver_diag.reason}. Correctness is unaffected — only speed.`
-                : "Array Block solve used the fast lattice-FFT coupling path (a regular same-height lattice of 16+ identical elements)."
-            }
-          >
-            <span>array path</span>
-            <span className="val">
-              {result.solver_diag.lattice_fft
-                ? "FFT lattice"
-                : result.solver_diag.operator === "HMatrix"
-                  ? "H-matrix"
-                  : "per-pair"}
-            </span>
-          </div>
-        )}
-        <div className="row">
-          <span>solve</span>
-          <span className="val">
-            {result ? `${result.solve_ms.toFixed(1)} ms` : "—"}
-          </span>
-        </div>
-        <div className="row">
-          <span>rtt</span>
-          <span className="val">
-            {rttMs != null ? `${rttMs.toFixed(1)} ms` : "—"}
-          </span>
-        </div>
-      </div>
-    </div>
-  );
-}
-
-function ViewPanel({
-  view,
-  size,
-  fill,
-  result,
-  preview,
-  sweep,
-  converge,
-  measured,
-  pattern,
-  pinnedPatterns,
-  measFreqMhz,
-  sweepRunning,
-  convergeRunning,
-  azElevDeg,
-  elevAzDeg,
-  cameraProjection,
-  showHeatmap,
-  showEnvelope,
-  showWireLabels = false,
-  showFeedNames = true,
-  multiFeed,
-  fineNorm,
-}: {
-  view: View;
-  size: number;
-  fill: boolean;
-  result: SolveResponse | null;
-  preview: SolveResponse | null;
-  sweep: SweepData | null;
-  converge: ConvergeData | null;
-  measured: MeasuredData | null;
-  pattern: PatternData | null;
-  pinnedPatterns: PinnedPattern[];
-  measFreqMhz: number;
-  sweepRunning: boolean;
-  convergeRunning: boolean;
-  azElevDeg: number;
-  elevAzDeg: number;
-  cameraProjection: Projection;
-  showHeatmap: boolean;
-  showEnvelope: boolean;
-  showWireLabels?: boolean;
-  showFeedNames?: boolean;
-  multiFeed: boolean;
-  fineNorm?: number | null;
-}) {
-  if (view === "antenna") {
-    // Fall back to the geometry-only preview while the real solve is in
-    // flight, but with the current heatmap/waveform overlays forced off —
-    // the preview has no currents, so only the bare wires + feed are drawn.
-    const showingPreview = !result && !!preview;
-    return (
-      <div className={fill ? "antenna-fill" : "antenna-thumb"}
-           style={fill ? undefined : { width: size, height: size }}>
-        <CurrentCanvas
-          result={result ?? preview}
-          projection={cameraProjection}
-          showHeatmap={showingPreview ? false : showHeatmap}
-          showEnvelope={showingPreview ? false : showEnvelope}
-          showWireLabels={showWireLabels}
-          showFeedNames={showFeedNames}
-          interactive={fill}
-        />
-      </div>
-    );
-  }
-  if (view === "azimuth") {
-    return (
-      <FarFieldChart
-        result={result}
-        pattern={pattern}
-        pinned={pinnedPatterns}
-        size={size}
-        cut="xy"
-        azElevDeg={azElevDeg}
-        elevAzDeg={elevAzDeg}
-        fineNorm={fineNorm}
-      />
-    );
-  }
-  if (view === "elevation") {
-    return (
-      <FarFieldChart
-        result={result}
-        pattern={pattern}
-        pinned={pinnedPatterns}
-        size={size}
-        cut="yz"
-        azElevDeg={azElevDeg}
-        elevAzDeg={elevAzDeg}
-        fineNorm={fineNorm}
-      />
-    );
-  }
-  return (
-    <SmithChart
-      r={result?.z_in_re ?? 0}
-      x={result?.z_in_im ?? 0}
-      z0={result?.z0_ohms ?? 50}
-      size={size}
-      sweep={sweep}
-      converge={converge}
-      measured={measured}
-      measFreqMhz={measFreqMhz}
-      running={sweepRunning}
-      convergeRunning={convergeRunning}
-      feeds={result?.feeds}
-      multiFeed={multiFeed}
-    />
-  );
-}
-
 // Fetch the scalar far-field metrics for a request (peak gain, takeoff, F/B,
 // beamwidths). Returns null when the design can't be evaluated or on error.
 async function fetchMetrics(req: SolveRequest): Promise<PatternMetrics | null> {
@@ -4615,115 +4186,4 @@ async function fetchMetrics(req: SolveRequest): Promise<PatternMetrics | null> {
   } catch {
     return null;
   }
-}
-
-function PatternCompareTable({
-  live,
-  liveLabel,
-  pinned,
-  onRemove,
-  onToggle,
-}: {
-  live: PatternMetrics | null;
-  liveLabel: string;
-  pinned: PinnedPattern[];
-  onRemove: (id: string) => void;
-  onToggle: (id: string) => void;
-}) {
-  const fmt = (v: number | undefined, d: number) =>
-    v === undefined || v === null ? "—" : v.toFixed(d);
-  // Live row's swatch reads the lobe CSS var so it matches the orange lobe in
-  // either theme; pinned rows use their fixed canvas ghost colors.
-  const rows = [
-    {
-      key: "live",
-      bg: "rgba(var(--plot-lobe-rgb), 0.95)",
-      label: liveLabel,
-      m: live,
-      enabled: true,
-      onToggle: undefined as undefined | (() => void),
-      onX: undefined as undefined | (() => void),
-    },
-    ...pinned.map((p) => {
-      const i = p.colorIdx % GHOST_COLOR_COUNT;
-      return {
-        key: p.id,
-        // CSS var (like the live row) so the swatch rethemes without a render.
-        bg: `rgba(var(--plot-ghost-${i}-rgb, ${GHOST_FALLBACK_RGB[i]}), 0.95)`,
-        label: p.label,
-        m: p.metrics,
-        enabled: p.enabled,
-        onToggle: () => onToggle(p.id),
-        onX: () => onRemove(p.id),
-      };
-    }),
-  ];
-  return (
-    <table className="compare-table">
-      <thead>
-        <tr>
-          <th>design</th>
-          <th>peak</th>
-          <th>takeoff</th>
-          <th>F/B</th>
-          <th>az bw</th>
-          <th />
-        </tr>
-      </thead>
-      <tbody>
-        {rows.map((row) => (
-          <tr key={row.key} className={row.enabled ? undefined : "compare-off"}>
-            <td className="compare-name">
-              {/* The whole swatch+name is the show/hide toggle — a big-enough
-                  touch target where the tiny swatch alone wouldn't be. The
-                  metrics stay readable while hidden; that's disable vs delete. */}
-              {row.onToggle ? (
-                <button
-                  type="button"
-                  className="compare-toggle"
-                  onClick={row.onToggle}
-                  aria-pressed={row.enabled}
-                  title={
-                    row.enabled
-                      ? "Hide this ghost overlay (keeps the pin)"
-                      : "Show this ghost overlay"
-                  }
-                >
-                  <span
-                    className="compare-swatch"
-                    style={{ background: row.bg }}
-                  />
-                  {row.label}
-                </button>
-              ) : (
-                <>
-                  <span
-                    className="compare-swatch"
-                    style={{ background: row.bg }}
-                  />
-                  {row.label}
-                </>
-              )}
-            </td>
-            <td>{fmt(row.m?.peak_gain_dbi, 1)}</td>
-            <td>{row.m ? `${fmt(row.m.takeoff_deg, 0)}°` : "—"}</td>
-            <td>{fmt(row.m?.front_to_back_db, 1)}</td>
-            <td>{row.m ? `${fmt(row.m.az_beamwidth_deg, 0)}°` : "—"}</td>
-            <td>
-              {row.onX && (
-                <button
-                  type="button"
-                  className="compare-x"
-                  onClick={row.onX}
-                  title="Remove this pinned pattern"
-                >
-                  ✕
-                </button>
-              )}
-            </td>
-          </tr>
-        ))}
-      </tbody>
-    </table>
-  );
 }
